refactor: move `app/env.ts`
